refactor(#4712): pure-move de-giant watchers/lifecycle.rs
